uncertainty-medical-imaging-1-main/requirements.txt
numpy
pyyaml
torch>=1.13
torchvision
scikit-learn
transformers
